split grouper and kernelize [pr] (#10854)
