+++ examples/openpilot/compile4.py
@@ -2,7 +2,7 @@ import sys, onnx
 from tinygrad import Tensor, fetch, GlobalCounters
 from tinygrad.uop.ops import UOp
 from tinygrad.frontend.onnx import OnnxRunner
-from tinygrad.engine.grouper import get_kernelize_map
+from tinygrad.engine.kernelize import get_kernelize_map
 from tinygrad.engine.schedule import create_schedule_with_vars
 from tinygrad.engine.realize import run_schedule
