+++ test/external/process_replay/process_replay.py
@@ -3,7 +3,7 @@
 import os, multiprocessing, logging, pickle, sqlite3, difflib, warnings, itertools
 from typing import Callable, Any
 from tinygrad.helpers import VERSION, Context, ContextVar, colored, db_connection, getenv, tqdm, to_function_name
-from tinygrad.engine.grouper import get_kernelize_map
+from tinygrad.engine.kernelize import get_kernelize_map
 from tinygrad.codegen.kernel import Kernel
 from tinygrad.uop.ops import UOp, Ops
