+++ tinygrad/tensor.py
@@ -14,7 +14,7 @@ from tinygrad.device import Device, Buffer
 from tinygrad.engine.realize import run_schedule
 from tinygrad.engine.memory import memory_planner
 from tinygrad.engine.schedule import ScheduleItem, create_schedule_with_vars
-from tinygrad.engine.grouper import get_kernelize_map
+from tinygrad.engine.kernelize import get_kernelize_map
 
 # *** all in scope Tensors are here. this gets relevant UOps ***
